Move bigquery test data and mock objects.
